+++ src/sentry/runner/commands/presenters/webhookpresenter.py
@@ -1,13 +1,14 @@
 from typing import Any, List, Optional, Tuple
 
+import requests
 from django.conf import settings
 
 from sentry import options
-from sentry.runner.commands.options.presenters.optionspresenter import OptionsPresenter
-from sentry.runner.commands.options.webhook import send_to_webhook
+from sentry.runner.commands.presenters.optionspresenter import OptionsPresenter
+from sentry.utils import json
 
 
-class SlackPresenter(OptionsPresenter):
+class WebhookPresenter(OptionsPresenter):
     """
     Sends changes of runtime options made via sentry configoptions
     to a webhook url in a truncated json format. The webhook url can
@@ -17,7 +18,8 @@ class SlackPresenter(OptionsPresenter):
 
     MAX_OPTION_VALUE_LENGTH = 30
 
-    def __init__(self) -> None:
+    def __init__(self, source: str) -> None:
+        self.source = source
         self.drifted_options: List[Tuple[str, Any]] = []
         self.channel_updated_options: List[str] = []
         self.updated_options: List[Tuple[str, Any, Any]] = []
@@ -28,7 +30,7 @@ class SlackPresenter(OptionsPresenter):
         self.invalid_type_options: List[Tuple[str, type, type]] = []
 
     @staticmethod
-    def is_slack_enabled():
+    def is_webhook_enabled():
         return (
             options.get("options_automator_slack_webhook_enabled")
             and settings.OPTIONS_AUTOMATOR_SLACK_WEBHOOK_URL
@@ -53,7 +55,7 @@ class SlackPresenter(OptionsPresenter):
 
         json_data = {
             "region": region,
-            "source": "options-automator",
+            "source": self.source,
             "drifted_options": [
                 {"option_name": key, "option_value": self.truncate_value(value)}
                 for key, value in self.drifted_options
@@ -80,7 +82,7 @@ class SlackPresenter(OptionsPresenter):
                 for key, got_type, expected_type in self.invalid_type_options
             ],
         }
-        send_to_webhook(json_data)
+        self._send_to_webhook(json_data)
 
     def truncate_value(self, value: str) -> str:
         value_str = str(value)
@@ -117,3 +119,12 @@ class SlackPresenter(OptionsPresenter):
         expected_type: type,
     ) -> None:
         self.invalid_type_options.append((key, got_type, expected_type))
+
+    def _send_to_webhook(self, json_data: dict) -> None:
+        if settings.OPTIONS_AUTOMATOR_SLACK_WEBHOOK_URL:
+            headers = {"Content-Type": "application/json"}
+            requests.post(
+                settings.OPTIONS_AUTOMATOR_SLACK_WEBHOOK_URL,
+                data=json.dumps(json_data),
+                headers=headers,
+            ).raise_for_status()
